feat: Reorder parts
