+++ .github/workflows/ci.yml
@@ -34,9 +34,8 @@ jobs:
           - "project/part3"
           - "project/part4"
           - "project/part5"
-          - "project/part6"
     steps:
-      - uses: actions/checkout@v5
+      - uses: actions/checkout@v6
 
       - uses: dtolnay/rust-toolchain@v1
         with:
