+++ project/part3/Cargo.toml
@@ -5,20 +5,14 @@ rust-version = "1.88"
 version      = "0.1.0"
 
 [dependencies]
-edge-captive = { version = "0.6.0", default-features = false, features = [
-  "io",
-] }
-edge-dhcp = "0.5.0"
-edge-nal = "0.5.0"
-edge-nal-embassy = "0.6.0"
 embassy-executor = "0.9.0"
 embassy-net = { version = "0.7.0", features = [
   "dhcpv4",
+  "dns",
   "medium-ethernet",
   "tcp",
   "udp",
 ] }
-embassy-sync = "0.7.0"
 embassy-time = "0.5.0"
 embedded-io-async = "0.6.1"
 esp-alloc = "0.9.0"
@@ -28,19 +22,24 @@ esp-backtrace = { version = "0.18.1", features = [
   "println",
 ] }
 esp-bootloader-esp-idf = { version = "0.4.0", features = ["esp32c3"] }
-esp-hal = { version = "1.0.0", features = ["esp32c3", "unstable"] }
+esp-hal = { version = "1.0.0", features = ["esp32c3", "log-04", "unstable"] }
 esp-println = { version = "0.16.1", features = ["esp32c3", "log-04"] }
-esp-radio = { version = "0.17.0", features = ["esp32c3", "unstable", "wifi"] }
-esp-rtos = { version = "0.2.0", features = ["embassy", "esp-radio", "esp32c3"] }
-heapless = { version = "0.8.0", features = ["serde"] }
-log = "0.4.28"
-edge-http = { version = "0.6.1", default-features = false, features = [
-  "edge-nal",
-  "embedded-io-async",
-  "io",
+esp-radio = { version = "0.17.0", features = [
+  "esp32c3",
+  "log-04",
+  "unstable",
+  "wifi",
 ] }
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-serde-json-core = "0.6.0"
+esp-rtos = { version = "0.2.0", features = [
+  "embassy",
+  "esp-radio",
+  "esp32c3",
+  "log-04",
+] }
+heapless = "0.9.1"
+log = "0.4.28"
+rust-mqtt = { version = "0.3.0", default-features = false }
+shtcx = { git = "https://github.com/Fristi/shtcx-rs", branch = "feature/async-support" }
 static_cell = "2.1.1"
 
 [profile.dev]
