+++ project/part3/src/mqtt.rs
@@ -1,10 +1,10 @@
 use core::fmt::Write;
 use embassy_net::{IpAddress, Ipv4Address, Stack, dns::DnsQueryType, tcp::TcpSocket};
-use embassy_time::{Duration as EmbassyDuration, Timer};
+use embassy_time::{Duration, Timer};
 use log::{debug, error, info};
 use rust_mqtt::{
     client::{client::MqttClient, client_config::ClientConfig as MqttClientConfig},
-    packet::v5::{publish_packet::QualityOfService, reason_codes::ReasonCode},
+    packet::v5::reason_codes::ReasonCode,
     utils::rng_generator::CountingRng,
 };
 
@@ -22,42 +22,37 @@ pub async fn mqtt_task(stack: Stack<'static>, mut sht: ShtC3<I2c<'static, esp_ha
 
     loop {
         // Wait for network to be ready before attempting connection
-        debug!("MQTT: Waiting for WiFi link to come up...");
+        debug!("Waiting for WiFi link to come up...");
         stack.wait_link_up().await;
-        debug!("MQTT: WiFi link up, waiting for network configuration...");
+        debug!("WiFi link up, waiting for network configuration...");
 
         // Wait for DHCP to assign an IP address
         stack.wait_config_up().await;
-        debug!("MQTT: Network configuration ready");
 
-        debug!("MQTT: Waiting to get IP address...");
+        debug!("Waiting to get IP address...");
         loop {
             if let Some(config) = stack.config_v4() {
-                debug!("MQTT: Got IP: {}", config.address);
+                debug!("Got IP: {}", config.address);
                 break;
             }
-            Timer::after(EmbassyDuration::from_millis(500)).await;
+            Timer::after(Duration::from_millis(500)).await;
         }
 
         // Check if we still have a valid network config before proceeding
         if !stack.is_config_up() {
-            debug!("MQTT: Network config lost, retrying...");
+            debug!("Network config lost, retrying...");
             continue;
         }
 
-        debug!("MQTT: Starting MQTT connection loop...");
-        Timer::after(EmbassyDuration::from_millis(1_000)).await;
+        Timer::after(Duration::from_millis(1_000)).await;
 
         let host = match BROKER_HOST {
-            Some(h) => {
-                debug!("MQTT: Using BROKER_HOST: {}", h);
-                h
-            }
+            Some(h) => h,
             None => {
-                debug!(
-                    "MQTT: No BROKER_HOST set. Provide e.g. BROKER_HOST=10.0.0.10 (or hostname) and optional BROKER_PORT."
+                error!(
+                    "No BROKER_HOST set. Provide e.g. BROKER_HOST=10.0.0.10 (or hostname) and optional BROKER_PORT."
                 );
-                Timer::after(EmbassyDuration::from_secs(5)).await;
+                Timer::after(Duration::from_secs(5)).await;
                 continue;
             }
         };
@@ -74,12 +69,12 @@ pub async fn mqtt_task(stack: Stack<'static>, mut sht: ShtC3<I2c<'static, esp_ha
                 Ok(addresses) if !addresses.is_empty() => addresses[0],
                 Ok(_) => {
                     error!("DNS query returned no addresses for {}", host);
-                    Timer::after(EmbassyDuration::from_secs(5)).await;
+                    Timer::after(Duration::from_secs(5)).await;
                     continue;
                 }
                 Err(e) => {
                     error!("DNS lookup error: {e:?}");
-                    Timer::after(EmbassyDuration::from_secs(5)).await;
+                    Timer::after(Duration::from_secs(5)).await;
                     continue;
                 }
             },
@@ -93,7 +88,7 @@ pub async fn mqtt_task(stack: Stack<'static>, mut sht: ShtC3<I2c<'static, esp_ha
         let connection = socket.connect(remote_endpoint).await;
         if let Err(e) = connection {
             error!("connect error: {:?}", e);
-            Timer::after(EmbassyDuration::from_secs(5)).await;
+            Timer::after(Duration::from_secs(5)).await;
             continue;
         }
         info!("connected!");
@@ -102,14 +97,14 @@ pub async fn mqtt_task(stack: Stack<'static>, mut sht: ShtC3<I2c<'static, esp_ha
             rust_mqtt::client::client_config::MqttVersion::MQTTv5,
             CountingRng(20000),
         );
-        config.add_max_subscribe_qos(QualityOfService::QoS1);
+        config.add_max_subscribe_qos(rust_mqtt::packet::v5::publish_packet::QualityOfService::QoS1);
         config.add_client_id("esp32c3");
         config.max_packet_size = 1024;
         let mut recv_buffer = [0; 512];
         let mut write_buffer = [0; 512];
+
         let write_len = write_buffer.len();
         let recv_len = recv_buffer.len();
-
         let mut client = MqttClient::<_, 5, _>::new(
             socket,
             &mut write_buffer,
@@ -119,37 +114,39 @@ pub async fn mqtt_task(stack: Stack<'static>, mut sht: ShtC3<I2c<'static, esp_ha
             config,
         );
 
-        if let Err(mqtt_error) = client.connect_to_broker().await {
-            match mqtt_error {
-                ReasonCode::NetworkError => error!("MQTT Network Error"),
-                _ => error!("Other MQTT Error: {:?}", mqtt_error),
-            }
-            continue;
+        match client.connect_to_broker().await {
+            Ok(()) => {}
+            Err(mqtt_error) => match mqtt_error {
+                ReasonCode::NetworkError => {
+                    error!("MQTT Network Error");
+                    continue;
+                }
+                _ => {
+                    error!("Other MQTT Error: {:?}", mqtt_error);
+                    continue;
+                }
+            },
         }
 
         // Main sensor reading and publishing loop
         loop {
             // Check network state before attempting operations
             if !stack.is_link_up() || !stack.is_config_up() {
-                debug!("MQTT: Network connection lost, reconnecting...");
+                debug!("Network connection lost, reconnecting...");
                 break;
             }
 
             // Read sensor
-            let (temp, humidity) = match read_sensor(&mut sht).await {
+            let (temp, _) = match read_sensor(&mut sht).await {
                 Some(reading) => reading,
                 None => {
-                    Timer::after(EmbassyDuration::from_secs(1)).await;
+                    Timer::after(Duration::from_secs(1)).await;
                     continue;
                 }
             };
 
-            // Format sensor values
-            let mut temperature_string = heapless::String::<32>::new();
+            let mut temperature_string: heapless::String<32> = heapless::String::new();
             write!(temperature_string, "{:.2}", temp).expect("write! failed!");
-
-            let mut humidity_string = heapless::String::<32>::new();
-            write!(humidity_string, "{:.2}", humidity).expect("write! failed!");
 
             // Helper to handle MQTT send errors
             let handle_mqtt_error = |mqtt_error: ReasonCode| match mqtt_error {
@@ -163,40 +160,27 @@ pub async fn mqtt_task(stack: Stack<'static>, mut sht: ShtC3<I2c<'static, esp_ha
                 }
             };
 
-            // Publish temperature
-            if let Err(e) = client
+            // MQTT
+            match client
                 .send_message(
                     "measurement/temperature",
                     temperature_string.as_bytes(),
-                    QualityOfService::QoS1,
+                    rust_mqtt::packet::v5::publish_packet::QualityOfService::QoS1,
                     true,
                 )
                 .await
             {
-                if handle_mqtt_error(e) {
-                    break; // Network error, reconnect
+                Ok(()) => {}
+                Err(mqtt_error) => {
+                    if handle_mqtt_error(mqtt_error) {
+                        break; // Network error, reconnect
+                    }
+                    continue;
                 }
-                continue;
-            }
-
-            // Publish humidity
-            if let Err(e) = client
-                .send_message(
-                    "measurement/humidity",
-                    humidity_string.as_bytes(),
-                    QualityOfService::QoS1,
-                    true,
-                )
-                .await
-            {
-                if handle_mqtt_error(e) {
-                    break; // Network error, reconnect
-                }
-                continue;
             }
 
             // Delay
-            Timer::after(EmbassyDuration::from_secs(1)).await;
+            Timer::after(Duration::from_secs(1)).await;
         }
     }
 }
