+++ project/part3/src/sensor.rs
@@ -1,17 +1,18 @@
-use embassy_time::{Duration as EmbassyDuration, Timer};
+use embassy_time::{Duration, Timer};
 use esp_hal::i2c::master::I2c;
 use log::{error, info};
 use shtcx::asynchronous::{PowerMode, ShtC3, max_measurement_duration};
 
 pub async fn read_sensor(sht: &mut ShtC3<I2c<'static, esp_hal::Async>>) -> Option<(f32, f32)> {
+    // Read sensor
     if let Err(e) = sht.start_measurement(PowerMode::NormalMode).await {
         error!("Failed to start measurement: {:?}", e);
+        Timer::after(Duration::from_secs(1)).await;
         return None;
     }
-
-    // Wait for measurement to complete
+    // Wait for 12.1 ms https://github.com/Fristi/shtcx-rs/blob/feature/async-support/src/asynchronous.rs#L413-L424
     let duration = max_measurement_duration(sht, PowerMode::NormalMode);
-    Timer::after(EmbassyDuration::from_micros(duration.into())).await;
+    Timer::after(Duration::from_micros(duration.into())).await;
 
     match sht.get_measurement_result().await {
         Ok(m) => {
