+++ project/part5/.cargo/config.toml
@@ -1,5 +1,5 @@
 [target.riscv32imc-unknown-none-elf]
-runner = "espflash flash --monitor --chip esp32c3"
+runner = "espflash flash --monitor --chip esp32c3 --partition-table partitions.csv"
 
 [env]
 ESP_LOG = "info"
