--- project/part6/.cargo/config.toml
@@ -1,26 +1,0 @@
-[target.riscv32imc-unknown-none-elf]
-runner = "espflash flash --monitor --chip esp32c3 --partition-table partitions.csv"
-
-[env]
-ESP_LOG = "info"
-
-[net]
-retry              = 3     # network retries
-git-fetch-with-cli = true  # use the `git` executable for git operations
-offline            = false
-
-[build]
-rustflags = [
-  # Required to obtain backtraces (e.g. when using the "esp-backtrace" crate.)
-  # NOTE: May negatively impact performance of produced code
-  "-C",
-  "force-frame-pointers",
-  # Link all sections
-  "-C",
-  "link-arg=-Tlinkall.x",
-]
-
-target = "riscv32imc-unknown-none-elf"
-
-[unstable]
-build-std = ["core"]
