--- project/part6/Cargo.toml
@@ -1,59 +1,0 @@
-[package]
-edition      = "2024"
-name         = "no_std-training"
-rust-version = "1.88"
-version      = "0.1.0"
-
-[dependencies]
-edge-captive = { version = "0.6.0", default-features = false, features = [
-  "io",
-] }
-edge-dhcp = "0.6.0"
-edge-http = { version = "0.6.1", default-features = false, features = [
-  "edge-nal",
-  "embedded-io-async",
-  "io",
-] }
-edge-nal = "0.5.0"
-edge-nal-embassy = "0.6.0"
-embassy-executor = "0.9.1"
-embassy-net = { version = "0.7.1", features = [
-  "dhcpv4",
-  "medium-ethernet",
-  "tcp",
-  "udp",
-] }
-embassy-sync = "0.7.2"
-embassy-time = "0.5.0"
-embedded-io-async = "0.6.1"
-embedded-storage = "0.3.1"
-esp-alloc = "0.9.0"
-esp-backtrace = { version = "0.18.1", features = [
-  "esp32c3",
-  "panic-handler",
-  "println",
-] }
-esp-bootloader-esp-idf = { version = "0.4.0", features = ["esp32c3"] }
-esp-hal = { version = "1.0.0", features = ["esp32c3", "unstable"] }
-esp-println = { version = "0.16.1", features = ["esp32c3", "log-04"] }
-esp-radio = { version = "0.17.0", features = ["esp32c3", "unstable", "wifi"] }
-esp-rtos = { version = "0.2.0", features = ["embassy", "esp-radio", "esp32c3"] }
-esp-storage = { version = "0.8.1", features = ["esp32c3"] }
-heapless = { version = "0.8.0", features = ["serde"] }
-log = "0.4.28"
-rust-mqtt = { version = "0.3.0", default-features = false }
-serde = { version = "1.0.228", default-features = false, features = ["derive"] }
-serde-json-core = "0.6.0"
-shtcx = { git = "https://github.com/Fristi/shtcx-rs", branch = "feature/async-support" }
-static_cell = "2.1.1"
-
-[profile.dev]
-# Rust debug is too slow.
-# For debug builds always builds with some optimization
-opt-level = "s"
-
-[profile.release]
-codegen-units = 1     # LLVM can perform better optimizations using a single thread
-debug         = 2
-lto           = 'fat'
-opt-level     = 's'
