--- project/part6/assets/templates/home.html
@@ -1,106 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset='utf-8'>
-    <meta name='viewport' content='width=device-width, initial-scale=1'>
-    <title>ESP WiFi Provisioning</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 40px;
-            background: #f0f0f0;
-        }
-
-        .container {
-            max-width: 600px;
-            margin: 0 auto;
-            background: white;
-            padding: 30px;
-            border-radius: 10px;
-            box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-        }
-
-        h1 {
-            color: #333;
-        }
-
-        .info {
-            background: #e3f2fd;
-            padding: 15px;
-            border-radius: 5px;
-            margin: 20px 0;
-        }
-
-        .status {
-            color: #4caf50;
-            font-weight: bold;
-        }
-
-        input,
-        button {
-            width: 100%;
-            padding: 12px;
-            margin: 8px 0;
-            box-sizing: border-box;
-            border: 1px solid #ddd;
-            border-radius: 4px;
-        }
-
-        button {
-            background: #2196F3;
-            color: white;
-            cursor: pointer;
-            font-weight: bold;
-        }
-
-        button:hover {
-            background: #0b7dda;
-        }
-    </style>
-</head>
-
-<body>
-    <div class='container'>
-        <h1>ESP32-C3 WiFi Provisioning</h1>
-        <div class='info'>
-            <p class='status'>+ Connected to ESP32-C3 Access Point</p>
-            <p>Device: <strong>esp-radio</strong></p>
-            <p>IP Address: <strong>192.168.2.1</strong></p>
-        </div>
-        <h2>Configure WiFi</h2>
-        <form id='wifi-form'>
-            <input type='text' id='ssid' placeholder='WiFi SSID' required>
-            <input type='password' id='password' placeholder='WiFi Password' required>
-            <button type='submit'>Save Configuration</button>
-        </form>
-        <script>
-            document.getElementById('wifi-form').addEventListener('submit', async function (e) {
-                e.preventDefault();
-                const ssid = document.getElementById('ssid').value;
-                const password = document.getElementById('password').value;
-
-                try {
-                    const response = await fetch('/save', {
-                        method: 'POST',
-                        headers: {
-                            'Content-Type': 'application/json',
-                        },
-                        body: JSON.stringify({ ssid, password })
-                    });
-
-                    if (response.ok) {
-                        window.location.href = '/saved';
-                    } else {
-                        alert('Failed to save configuration');
-                    }
-                } catch (error) {
-                    alert('Error: ' + error.message);
-                }
-            });
-        </script>
-        <p style='text-align: center; color: #999; margin-top: 20px;'>Powered by Rust & Embassy</p>
-    </div>
-</body>
-
-</html>
