--- project/part6/assets/templates/saved.html
@@ -1,47 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset='utf-8'>
-    <meta name='viewport' content='width=device-width, initial-scale=1'>
-    <title>Configuration Saved</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 40px;
-            background: #f0f0f0;
-        }
-
-        .container {
-            max-width: 600px;
-            margin: 0 auto;
-            background: white;
-            padding: 30px;
-            border-radius: 10px;
-            box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-        }
-
-        h1 {
-            color: #4caf50;
-        }
-
-        .success {
-            background: #c8e6c9;
-            padding: 15px;
-            border-radius: 5px;
-            margin: 20px 0;
-        }
-    </style>
-</head>
-
-<body>
-    <div class='container'>
-        <h1>Configuration Saved!</h1>
-        <div class='success'>
-            <p>Your WiFi credentials have been received.</p>
-            <p>Check the serial console for details.</p>
-        </div>
-    </div>
-</body>
-
-</html>
