--- project/part6/rust-toolchain.toml
@@ -1,4 +1,0 @@
-[toolchain]
-channel    = "stable"
-components = ["rust-src"]
-targets = ["riscv32imc-unknown-none-elf"]
